Authenticación con ID y contraseñas encriptadas en base de datos

+++ Private Version/Server/src/ar/edu/unicen/isistan/si/soploon/server/models/Correction.java
@@ -7,9 +7,9 @@ import com.google.gson.annotations.Expose;
 public class Correction {
 
 	@Expose
-	private int userId;
+	private long userId;
 	@Expose
-	private int projectId;
+	private long projectId;
 	@Expose
 	private long date;
 	@Expose
@@ -19,19 +19,19 @@ public class Correction {
 	@Expose
 	private Configuration configuration;
 
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
 
-	public int getProjectId() {
+	public long getProjectId() {
 		return projectId;
 	}
 
-	public void setProjectId(int projectId) {
+	public void setProjectId(long projectId) {
 		this.projectId = projectId;
 	}
